apps: rpc_demod: change copy memory from RPMsg buffer
As the memory form RPMsg buffer can be UIO memory on Linux side,
we should the memory copy operation in this example.

In future, the proper fix should be the rx callback should also
return the memory I/O region of the RPMsg buffer, and it should
use the I/O region memory operation for memory copy.

Signed-off-by: Wendy Liang <jliang@xilinx.com>