rpmsg: rpmsg_send returns number of bytes has sent
This patch change rpmsg_send return value from just return
if it has sent data to return the number of bytes of data has
sent or negative value on failure.

There is case that the input data is larger than the max RPMsg
buffer size, in this case, rpmsg_send will just send the max
RPMsg buffer size data. However, the user will not know if
we don't return the size of data it has actually sent.

Signed-off-by: Wendy Liang <jliang@xilinx.com>