linux remoteproc: not use "-1" as page shift for vring I/O
As the da address in the resource table is the offset instead of
the actual physical address, we need to have another I/O with the
shared memory to da mapping for vrings.

When creating such a metal I/O region, we should not use "-1"
as page_shift, instead use the value should be based on the
actual size of the physical address field as the "da" address
is flat.

Signed-off-by: Wendy Liang <jliang@xilinx.com>