apps: linux userspace: rsc table: add get rsc table to API to header file
Add get resource table function to header file so that other file
can see it.

As the resource table can be updated in the OpenAMP library, remove
"const" attribute of the resource table.

In using OpenAMP in Linux userspace case, the rsc table is not yet
put into the shared memory, and thus, updating it doesn't have
impact on the other side. This will need to change.

Signed-off-by: Wendy Liang <jliang@xilinx.com>