apps: linux userspace: rsc table: set VDEV status to DRVIER_OK
As when using OpenAMP in Linux userspace, the resource table
is not in shared memory, vdev status in resource table updated
in master side will not be noted by the slave side. We need to
use static settings in the resource table and make the vdev
status to DRIVER_OK.

This will need to fix.

Signed-off-by: Wendy Liang <jliang@xilinx.com>