apps: linux userspace: set vdev IPI
We need to set vdev IPI for vdev configuration change notification.
Even for now, in Linux userspace, the vdev master/slave doesn't
exchange vdev changes as the vdev configuration and status information
is in the resource table, and the resource table is not in shared
memory, it is required by the library.

And this will change, the resource table should be put in the shared
memory.

Signed-off-by: Wendy Liang <jliang@xilinx.com>