virtqueue: encapsulate shared memory handling
Encapsulate the the handling of the shared memory region to make it clear
exactly what functionality is needed and ensure that others aren't touch
internal data that we aren't aware of.

We treat the sh_mem region in the virtqueue as an obaque pointer that
only should get used by v2p and p2v functions.

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>