rpmsg: rpmsg_ns_callback: set ept dest addr from ns_msg.addr
The src address in the rpmsg header might be bogus (ie the sender might
not have actually set the src address explicitly so we can't trust it).
We should use the address in the ns_msg to set the dest_addr of the
endpoint.

TODO: We might want to move this so its done before we call
new_endpoint_cb() in the case we created the endpoint.

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>