remoteproc: elf: loader returns what's need to load only
Instead of also loading the required data, ELF loader will just
parse the headers and returns what's the next to load.
The remoteproc load() will be responsible to get the data from
image file or load data to the target memory based on the returned
address information from the loader.

Signed-off-by: Wendy Liang <jliang@xilinx.com>