apps: remove detecting shutdown message.
Shutdown message is not a standard RPMsg message, and it is always
required. It is up to the application to decide whether they want
to notify the remote that it is going to shutdown.

Remove sending shutdown message from the demos. When destroying
the endpoint, it will send out service destroy message if name
service is enabled.

Signed-off-by: Wendy Liang <jliang@xilinx.com>