rpmsg: Make RPMSG_BUFFER_SIZE a fixed define
Since the rpmsg_virtio.h file is exported, it was possible to have
different values for RPMSG_BUFFER_SIZE when open-amp is built and
when the library is included in an application.
By making this parameter a CMake option and hard-coding the value
in rpmsg_virtio.h, it will make sure this parameter cannot differ
between building of open-amp and the user application.

Signed-off-by: Gaute Nilsson <gaute.nilsson@siemens-energy.com>