rpmsg: Introduce flow control channel
Create a dedicated channel for the endpoints flow control.
This channel implements the software control allowing to:
- inform the remote side that it is ready or not to communicate,
- be informed by the remote endpoint that it can start or must
  stop to send rpmsg.

In addition the channel can help to bind the local and the remote
endpoint after a NS announcement as it set the local endpoint
dest_addr if is current value is RPMSG_ADDR_ANY.

To enable the feature the VIRTIO_RPMSG_F_FC must be set in the
vdev structure in the resource table.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>