virtio: Use physical addresses for buffers in virtio
Have the higher level rpmsg_virtio handle the virt-to-phys and back
conversion. The rpmsg_virtio layer already has the shared buffer memory
metal IO. This removes the need to pass in the buffer region metal_io
struct down to the virtio layer, which otherwise does nothing with the
contents of the buffers.

Signed-off-by: Andrew Davis <afd@ti.com>