virtqueue: Fix comment on shm_io and fix type
This should hold a pointer to an metal_io_region, make that the type.

Also fix the comment, this region holds the address of the message
buffers, not the vring's descriptor table nor available/used rings.
It is only used for virt-to-phys/phys-to-vert on the buffers pointed
to by these descriptors.

This comment seems to have cause an issue in virtio_mmio_drv where
this region was used to translate the address of the vring descriptors.
This may have worked if the vring descriptors where part of the same
IO space as the buffers they point to, but this is not guaranteed
to always be the case. Fix that here.

Signed-off-by: Andrew Davis <afd@ti.com>