virtqueue: Fix comment on shm_io and fix type
This should hold a pointer to a metal_io_region, make that the type.

Also fix the comment above this variable. This region holds the address
of the message buffers, not the vring descriptor table nor available/used
ring data. It is only used for virt-to-phys/phys-to-vert translation on
the buffers pointed to by these descriptors.

This comment seems to have cause an issue in virtio_mmio_drv where
this region is used to translate the address of the vring descriptor
table. This may have worked if the vring descriptor table was part of
the same IO space as the buffers point to by those descriptors, but
this is not guaranteed to always be the case. Fix that here.

Signed-off-by: Andrew Davis <afd@ti.com>