virtio-mmio: Use the vmdev IO region for vq_ring.desc
The metal_io_region in the virtqueue struct is for translating
the buffer address stored in the virtqueue descriptor table, not
the address of the descriptor table itself.

This may have worked previously if the vring descriptors where part
of the same IO space as the buffers they point to, but this is not
guaranteed to always be the case. Fix that here.

Signed-off-by: Andrew Davis <afd@ti.com>