CI: Fix compliance check folder bypass
On GitHub, this check is tagged as "Required."
However, if a PR only affects the following folders:.github, docs,
scripts, or cmake, the check is not executed (to avoid false positives),
resulting in the status "execution pending" + "required" on GitHub.

This commit forces the execution of the check but bypasses the
compliance.xml test when the PR only affects the listed folders.

If at least one update affects other folders, the compliance check
runs as usual.

In addition, the cmake folder is removed from the list, as there is
no reason to skip compliance tests on it.

Signed-off-by: Arnaud Pouliquen arnaud.pouliquen@foss.st.com